+++ RazorSvelte/App/404.svelte
@@ -1,12 +1,20 @@
 <script lang="ts">
-    import Layout from "./shared/layout.svelte";
+    import Layout from "./shared/layout/main.svelte";
 </script>
 
 <Layout>
-    <div class="bx--content bx--toast-notification--error">
-        <h1 class="">404 Not Found</h1>
+    <div class="container text-center">
+        <h1 class="text-danger">404 Not Found</h1>
+        
         <p>
             This page doesn't seem to exists.
         </p>
     </div>
 </Layout>
+
+<style lang="scss">
+    :global(main) {
+        margin: auto;
+    }
+</style>
+