+++ RazorSvelte/App/authorized-page.svelte
@@ -1,20 +1,33 @@
 <script lang="ts">
-    import Layout from "./shared/layout.svelte";
-    import { UnorderedList, ListItem, } from "carbon-components-svelte";
+    import Layout from "./shared/layout/main.svelte";
     import { getFromJson } from "./shared/config";
 
     let user = getFromJson<{name: string, email: string, timezone: string, timestamp: string}>("user");
 </script>
 
 <Layout>
-    <div class="bx--content bx--toast-notification--info">
+    <div class="container text-center m-5">
         <h1>Authorized Access Page</h1>
-        
-        <UnorderedList expressive>
-            <ListItem>Name: {user.name}</ListItem>
-            <ListItem>Email: {user.email}</ListItem>
-            <ListItem>Timezone: {user.timezone}</ListItem>
-            <ListItem>Timestamp: {user.timestamp}</ListItem>
-        </UnorderedList>
+        <ul class="list-group list-group-flush">
+            <li class="list-group-item">
+                Name: {user.name}
+            </li>
+            <li class="list-group-item">
+                Email: {user.email}
+            </li>
+            <li class="list-group-item">
+                Timezone: {user.timezone}
+            </li>
+            <li class="list-group-item">
+                Timestamp: {user.timestamp}
+            </li>
+        </ul>
     </div>
 </Layout>
+
+<style lang="scss">
+    :global(main) {
+        margin: auto;
+    }
+</style>
+