+++ RazorSvelte/App/index.svelte
@@ -1,96 +1,74 @@
 <script lang="ts">
-    import Layout from "./shared/layout.svelte";
-    import { 
-        Select, 
-        SelectItem, 
-    
-        DataTable,
-        Toolbar,
-        ToolbarContent,
-        ToolbarSearch,
-        ToolbarMenu,
-        ToolbarMenuItem,
-        Pagination,
+    import Layout from "./shared/layout/main.svelte";
+    import Modal from "./shared/components/modal.svelte";
+    import type ModalOptions from "./shared/components/modal";
 
-        Button } from "carbon-components-svelte";
+    let selectionValue: string = "";
+    let selectionElement: HTMLSelectElement;
+    $: selectedText = selectionValue ? selectionElement.options[selectionElement.selectedIndex].innerText : "";
 
-    let items = [
-            {name: "ASP.NET", url: "https://docs.microsoft.com/aspnet/"},
-            {name: ".NET", url: "https://docs.microsoft.com/dotnet/"},
-            {name: "C#", url: "https://docs.microsoft.com/en-us/learn/paths/build-dotnet-applications-csharp/"},
-            {name: "Svelte", url: "https://svelte.dev/"},
-            {name: "Svelte Tutorial", url: "https://svelte.dev/tutorial/basics"},        
-            {name: "Svelte Carbon Components", url: "https://carbon-components-svelte.onrender.com/"},
-            {name: "Svelte Carbon Components Source", url: "https://github.com/carbon-design-system/carbon-components-svelte"},
-            {name: "Svelte Carbon Components Icons", url: "https://github.com/carbon-design-system/carbon-icons-svelte"},
-            {name: "Carbon Charts", url: "https://github.com/carbon-design-system/carbon-charts"},
-            {name: "Carbon Design System", url: "https://www.carbondesignsystem.com/"},
-            {name: "Typescript", url: "https://www.typescriptlang.org/"},
-            {name: "SASS", url: "https://sass-lang.com/guide"},
-    ];
+    let modal1: ModalOptions = {
+        open: false,
+        content: "some content",
+        title: "some title",
+        titleCloseButton: true,
+        closeBtn: true,
+        large: true,
+        options: {backdrop: true, focus: true, keyboard: true},
+        buttons: [
+            {text: "Do", click: () => modal1.open = false}
+        ]
+    };
 
-    let selected = "5";
-
-    let rows = Array.from({ length: 10 }).map((_, i) => ({
-        id: i,
-        name: "Load Balancer " + (i + 1),
-        protocol: "HTTP",
-        port: 3000 + i * 10,
-        rule: i % 2 ? "Round robin" : "DNS delegation",
-    }));
-
-    let pagination = {pageSize: 10, page: 0}
+    let modal2 = {open: false, closeBtn: true}
 </script>
 
 <Layout>
-    <Select inline light labelText="Choose what would you like to learn today" bind:selected>
-        {#each items as item, i}
-            <SelectItem value={i.toString()} text={item.name}></SelectItem>
-        {/each}
-    </Select>
+    <div class="container pt-4">
 
-    <Button href="{items[Number(selected)].url}" target="_target">
-        Learn about {items[Number(selected)].name}
-    </Button>
+        <h1 class="text-center text-primary">
+            Hello World from Svelte, Boostrap and Razor
+        </h1>
 
-    <br />
-    <br />
-    <hr />
-    <br />
+        <div class="shadow-lg card mt-3">
+            <div class="card-body">
+                <div class="card-title h5">Select What Do You Want to Learn Today</div>
+                <select class="form-select form-select-lg mb-3" bind:value={selectionValue} bind:this={selectionElement}>
+                    <option selected value="">Open this select menu</option>
+                    <option value="https://docs.microsoft.com/aspnet/core">Building Web apps with ASP.NET Core</option>
+                    <option value="https://svelte.dev/">SVELTE: CYBERNETICALLY ENHANCED WEB APPS</option>
+                    <option value="https://svelte.dev/tutorial/basics">Svlete Tutorial</option>
+                    <option value="https://getbootstrap.com/docs/5.1/getting-started/introduction/">Bootstrap</option>
+                    <option value="https://sass-lang.com/guide">SASS and SCSS Languague</option>
+                </select>
+                {#if selectionValue}
+                <a class="btn btn-primary" href="{selectionValue}" target="_blank">{selectedText}</a>
+                {/if}
+            </div>
+        </div>
 
-    <DataTable sortable title="Load balancers" description="Your organization's active load balancers."
-        headers={[
-            { key: "name", value: "Name" },
-            { key: "protocol", value: "Protocol" },
-            { key: "port", value: "Port" },
-            { key: "rule", value: "Rule" },
-        ]}
-        pageSize={pagination.pageSize}
-        page={pagination.page}
-        {rows}>
-        <Toolbar>
-            <ToolbarContent>
-                <ToolbarSearch persistent value="round" shouldFilterRows={(row, value) => {
-                return (
-                    /(6|8)$/.test(row.name) && row.rule.toLowerCase().includes(value.toString().toLowerCase())
-                );}}
-            />
-            <ToolbarMenu>
-                <ToolbarMenuItem primaryFocus>Restart all</ToolbarMenuItem>
-                <ToolbarMenuItem href="https://cloud.ibm.com/docs/loadbalancer-service">
-                    API documentation
-                </ToolbarMenuItem>
-                <ToolbarMenuItem hasDivider danger>Stop all</ToolbarMenuItem>
-            </ToolbarMenu>
-            <Button>Create balancer</Button>
-            </ToolbarContent>
-        </Toolbar>
-    </DataTable>
-    <Pagination
-        bind:pageSize={pagination.pageSize}
-        bind:page={pagination.page}
-        totalItems={6}
-        pageSizeInputDisabled
-    />
+        <div class="shadow-lg card mt-3">
+            <div class="card-body">
+                <div class="card-title h5">Modal Component Demo</div>
+                <div>
+                    <button class="btn btn-secondary" on:click={() => modal1.open = true}>Examle 1</button>
+                    <button class="btn btn-secondary" on:click={() => modal2.open = true}>Example 2</button>
+                </div>
+            </div>
+        </div>
 
+    </div>
 </Layout>
+
+<Modal options={modal1} />
+
+<Modal options={modal2}>
+    <span slot="title">modal2 title</span>
+
+    <div class="text-center">
+        <i class="spinner-border" style="width: 3rem; height: 3rem;"></i>
+    </div>
+
+    <span slot="footer">modal2 footer</span>
+</Modal>
+