+++ RazorSvelte/App/shared/layout/footer.svelte
@@ -1,0 +1,11 @@
+﻿<script lang="ts">
+    import urls from "../urls";
+</script>
+
+<footer class="footer mt-auto py-3 bg-primary bg-opacity-10">
+    <div class="container py-5">
+        <span class="text-muted">
+            &copy; 2021 - RazorSvelte - <a href="{urls.privacyUrl}">Privacy</a>
+        </span>
+    </div>
+</footer>