+++ RazorSvelte/App/shared/layout/main.svelte
@@ -1,0 +1,10 @@
+<script lang="ts">
+    import Header from "./header.svelte";
+    import Footer from "./footer.svelte";
+</script>
+
+<Header />
+<main>
+    <slot></slot>
+</main>
+<Footer />