+++ RazorSvelte/App/spa-example.svelte
@@ -1,52 +1,56 @@
 <script lang="ts">
-    import Layout from "./shared/layout.svelte";
-    import { Tile, Tabs, Tab } from "carbon-components-svelte";
-
-
+    import Layout from "./shared/layout/main.svelte";
     import Router from "svelte-spa-router";
     import Home from "./spa-views/home.svelte";
     import About from "./spa-views/about.svelte";
     import Parametrized from "./spa-views/parametrized.svelte";
     import Unknown from "./spa-views/unknown.svelte";
+
+    let hash = document.location.hash;
+    window.onhashchange = () => hash = document.location.hash;
 </script>
 
 <Layout>
-    <Tile>
-        <h1>SPA (Single Page Application) Example</h1>
-        <p>
-            Single Page Application or SPA have page router on the client, rather then on server.
-        </p>
-        <p>
-            This example uses svelte component called <code style="font-weight: bold;">svelte-spa-router</code> 
-            that implements client-side router by using hashtags.
-            You can use any other client router implementation or library or implement it youserlf it is not that hard.
-        </p>
-        <p>
-            <a href="https://www.npmjs.com/package/svelte-spa-router">svelte-spa-router GitHub</a>
-        </p>
-        <p>
-            <a class="" href="https://svelte.dev/repl/6e7546119cff41fabe35df3ef999ff7e?version=3.29.4">svelte-spa-router REPL</a>
-        </p>
-        <p>
-            Bellow is implement where each SPA view or page is inside each tab component.
-        </p>
-    </Tile>
+    <div class="container mt-3">
 
-    <br />
+        <h3>Spa Example</h3>
+        <p>
+            This is an example of the Single-Page App (SPA) by using 
+            <a class="btn-link" href="https://www.npmjs.com/package/svelte-spa-router">svelte-spa-router</a> component.
+        </p>
+        <p>
+            See <a class="btn-link" href="https://svelte.dev/repl/6e7546119cff41fabe35df3ef999ff7e?version=3.29.4">REPL</a> and/or 
+            <a class="btn-link" href="https://github.com/ItalyPaleAle/svelte-spa-router">GitHub</a> for this component.
+        </p>
 
-    <Tabs autoWidth>
-        <Tab label="Home" on:click={() => document.location = "#"} />
-        <Tab label="About" on:click={() => document.location = "#/about"} />
-        <Tab label="Parametrized with AAA" on:click={() => document.location = "#/parametrized/AAA"} />
-        <Tab label="Parametrized with BBB" on:click={() => document.location = "#/parametrized/BBB"} />
-        <Tab label="Unknown route test" on:click={() => document.location = "#/test"} />
-    </Tabs>
+        <ul class="nav nav-tabs">
+            <li class="nav-item">
+                <!-- svelte-ignore a11y-invalid-attribute -->
+                <a class="nav-link" class:active={hash == ""} href="#">Home</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" class:active={hash == "#/about"} href="#/about">About</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" class:active={hash == "#/parametrized/AAA"} href="#/parametrized/AAA">Parametrized with "AAA"</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" class:active={hash == "#/parametrized/BBB"} href="#/parametrized/BBB">Parametrized with "BBB"</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="#/test">Unknown route</a>
+            </li>
+        </ul>
 
-    <Router routes={{
-        "/": Home,
-        "/about": About,
-        "/parametrized/:param": Parametrized,
-        "*": Unknown,
-    }} />
+        <div class="p-0 mt-3">
+            <Router routes={{
+                "/": Home,
+                "/about": About,
+                "/parametrized/:param": Parametrized,
+                "*": Unknown,
+                //'/blog/:blogName': Blog
+            }} />
+        </div>
+    </div>
 </Layout>