+++ RazorSvelte/Pages/About.entry.ts
@@ -1,3 +1,0 @@
-/// <reference types="svelte" />
-import App from "../App/about.svelte";
-export default new App({target: document.body});