+++ RazorSvelte/Pages/About.rollup.config.js
@@ -1,3 +1,0 @@
-import config from "../Scripts/rollup.config";
-export default config("./Pages/About.entry.ts");
-