+++ RazorSvelte/Pages/AuthorizedPage.rollup.config.js
@@ -1,2 +1,2 @@
 import config from "../Scripts/rollup.config";
-export default config({ input: "./Pages/AuthorizedPage.entry.ts", jsOutput: "./wwwroot/build/authorized-page.js", cssOutput: "authorized-page.css" });
+export default config("./Pages/AuthorizedPage.entry.ts");