+++ RazorSvelte/Pages/Index.entry.ts
@@ -1,6 +1,3 @@
 /// <reference types="svelte" />
 import App from "../App/index.svelte";
-
-export default new App({
-    target: document.body
-});
+export default new App({target: document.body});