+++ RazorSvelte/Pages/Privacy.entry.ts
@@ -1,0 +1,3 @@
+/// <reference types="svelte" />
+import App from "../App/privacy.svelte";
+export default new App({ target: document.body });