+++ RazorSvelte/Pages/SpaExample.rollup.config.js
@@ -1,2 +1,2 @@
 import config from "../Scripts/rollup.config";
-export default config({ input: "./Pages/SpaExample.entry.ts", jsOutput: "./wwwroot/build/spa-example.js", cssOutput: "spa-example.css" });
+export default config("./Pages/SpaExample.entry.ts");