+++ RazorSvelte/Scripts/rollup.config.js
@@ -6,7 +6,6 @@ import sveltePreprocess from "svelte-preprocess";
 import typescript from "@rollup/plugin-typescript";
 import css from "rollup-plugin-css-only";
 import replace from '@rollup/plugin-replace';
-import { optimizeImports } from "carbon-preprocess-svelte";
 
 const production = !process.env.ROLLUP_WATCH;
 
@@ -48,22 +47,16 @@ export default (param, globals) => {
                 "process.env.NODE_ENV": JSON.stringify("development")
             }),
             svelte({
-                preprocess: [sveltePreprocess({ sourceMap: !production }), optimizeImports()],
+                preprocess: sveltePreprocess({ sourceMap: !production }),
                 compilerOptions: {
                     // enable run-time checks when not in production
                     dev: !production
-                },
-                onwarn: (warning, handler) => {
-                    const { code } = warning;
-                    // skip css unused selector warnings
-                    if (code === "css-unused-selector") {
-                        return;
-                    }
-                    handler(warning);
-                },
+                }
             }),
+            // we"ll extract any component CSS out into
+            // a separate file - better for performance
             css({ output: cssOutput }),
-    
+
             // If you have external dependencies installed from
             // npm, you"ll most likely need these plugins. In
             // some cases you"ll need additional configuration -